Add order/index route and template
